+++ core/services/src/main/java/pl/daniel/services/MoviesService.kt
@@ -7,6 +7,6 @@ import retrofit2.http.Query
 interface MoviesService {
 
     @GET("movie/now_playing")
-    suspend fun nowPlaying(@Query("page") page: Int,): NowPlayingDto
+    suspend fun nowPlaying(@Query("page") page: Long): NowPlayingDto
 
 }
